requirements.txt
flask==2.2.5
openai
twilio
requests
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
elevenlabs
google-auth-httplib2
